Move shared Proxmox helpers
